requirements.txt
# STPA - Requirements
# Install with: pip install -r requirements.txt

# Core scientific stack
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Dashboard
flask>=3.0.0
requests>=2.31.0

# Visualization
plotly>=5.17.0

# PDF handling (install at least one)
pikepdf>=8.0.0
PyPDF2>=3.0.0
pdfminer.six>=20221105

# Data loading
openpyxl>=3.1.0   # For Excel data files

# Dev tools
pytest>=7.4.0
httpx>=0.25.0     # For FastAPI test client
